nested try catch
